+++ aerosense/__init__.py
@@ -13,7 +13,7 @@ Usage:
 from typing import List
 
 # --- Versioning ---
-__version__: str = "v5.6.0"
+__version__: str = "v5.7.0"
 __release__: str = "2026-03-25"
 
 # --- Component Exports ---